+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - `authService.signUp()`
 
+## [4.0.0] - 2023-05-26
+
 ### Changed
 
 - **[BREAKING]** `formService.getFormElementDynamicOptions()` function result has changed to accommodate options sets that support passing a querystring parameter to filter options server side